Fix Convenience Kinetics Exchange for >1 Reaction
In convenience kinetics, exchange was set using the equality operator,
which led to the last reaction to update the exchange of metabolite
overwriting all previous ones. This commit fixes this to accumulate
exchange updates across reactions.

This bug led to both internal and external antibiotic concentrations
rising in the antibiotics process, which this commit fixes.