Create Tools to Save, Load, & Check Reference Data
For many processes, it is difficult to write simple assertions. For
example, it is not easy to write an assertion that a growth curve is as
expected. Instead of explicitly describing the curves we expect (e.g. as
an equation), we can save reference data that describes expected
behavior and then compare output to that reference. This commit
introduces the tools to:

* Save reference timeseries as CSV files
* Load reference timeseries from CSV files
* Compare timeseries to check that they are sufficiently correlated. We
  check for correlation instead of equality to allow for handling of
  stochastic processes.