Generalize Toy Antibiotic Injector Process
For testing the death process, I created a toy injector process to
inject antibiotic into the cell. I also need to inject nucleotides into
the cell for the antibiotic resistance operon pending integration of
metabolism, so this commit creats a generic injector process that can
serve both roles. The injectors in the death process are replaced with
this generic one.