Fix Growth-Division Experiment Environment Type
The growth and division experiment was failing because the
growth_division composite includes a glucose-lactose transport process,
but the experiment was using a lattice environment which uses minimal
media without the necessary metabolites.

In particular, the error was:

    KeyError: 'glc__D_e'

This commit fixes this issue by switching the growth and division
experiment to use ecoli_core_glc.