Move File-Dependent Lines from Module to __init__
Before, when the flagella chromosome or complexation modules were
imported, the fasta file would be read because those modules called
read_sequence at the top of the file. This prevented autodoc from
inspecting those files for docstrings. This commit moves the calls to
read_sequence into constructors so that the code can be imported even if
the fasta file is unavailable.