Mock or Import Vivarium Dependencies for Docs
Now that we are using autodoc to compile the reference documentation, we
need to make all the Vivarium code importable. We do this by mocking all
the dependencies we can and importing in the docs requirements.txt file
any dependencies we can't mock. This is required for Read the Docs to be
able to build the documentation.