Load E. coli Chromosome from Absolute Path
The E. coli chromosome fasta file was being loaded using a path relative
to the root. This caused problems whenever the file was loaded on import
because autodoc imports files from the /doc directory. This commit
changes the flagella chromosome to compute an absolute path to the fasta
file using the chromsome python file as a reference point.